feat: add more functionalities

+++ Sources/PluginEngine/Engine/Utils/FileUtils.swift
@@ -10,9 +10,51 @@ import Foundation
 import PluginInterface
 
 public class FileUtils: FileUtilsProtocol {
-    private(set) var currentDir: URL?
+    public private(set) var currentWorkSpace: URL?
+    
+    public var currentWorkSpacePath: String? {
+        get {
+            currentWorkSpace?.absoluteFilePath
+        }
+    }
     
     public init() {}
+    
+    public func openFile(at path: String) throws -> Data {
+        guard let currentDir = currentWorkSpace else {
+            throw FileUtilsErrors.noSelectedDir
+        }
+        let fileURL = currentDir.appending(component: path)
+        if !fileURL.isFileURL {
+            throw FileUtilsErrors.invalidFileURL(url: fileURL)
+        }
+        
+        return try Data(contentsOf: fileURL)
+    }
+    
+    public func writeFile(at path: String, with content: String) throws {
+        guard let currentDir = currentWorkSpace else {
+            throw FileUtilsErrors.noSelectedDir
+        }
+        
+        let fileURL = currentDir.appending(component: path)
+        if !fileURL.isFileURL {
+            throw FileUtilsErrors.invalidFileURL(url: fileURL)
+        }
+        try self.writeFile(at: fileURL, with: content)
+    }
+    
+    public func delete(at path: String) throws {
+        guard let currentDir = currentWorkSpace else {
+            throw FileUtilsErrors.noSelectedDir
+        }
+        
+        let fileURL = currentDir.appending(component: path)
+        if !fileURL.isFileURL {
+            throw FileUtilsErrors.invalidFileURL(url: fileURL)
+        }
+        try self.delete(at: fileURL)
+    }
 
     public func updateCurrentWorkSpace() throws -> URL {
         let dialog = NSOpenPanel()
@@ -22,7 +64,7 @@ public class FileUtils: FileUtilsProtocol {
 
         if dialog.runModal() == .OK {
             if let result = dialog.url {
-                currentDir = result
+                currentWorkSpace = result
                 return result
             }
         }
@@ -31,30 +73,44 @@ public class FileUtils: FileUtilsProtocol {
     }
 
     public func list() throws -> [String] {
-        guard let currentDir = currentDir else {
+        guard let currentWorkSpace = currentWorkSpace else {
             throw FileUtilsErrors.noSelectedDir
         }
 
         let fm = FileManager.default
-        let path = currentDir.absoluteString.replacingOccurrences(of: "file://", with: "")
+        guard let path = currentWorkSpace.absoluteFilePath else {
+            throw FileUtilsErrors.invalidFileURL(url: currentWorkSpace)
+        }
         let items = try fm.contentsOfDirectory(atPath: path)
         return items
     }
 
     public func openFile(at path: URL) throws -> Data {
+        if !path.isFileURL {
+            throw FileUtilsErrors.invalidFileURL(url: path)
+        }
         return try Data(contentsOf: path)
     }
 
     public func writeFile(at path: URL, with content: String) throws {
+        if !path.isFileURL {
+            throw FileUtilsErrors.invalidFileURL(url: path)
+        }
         try content.write(to: path, atomically: true, encoding: .utf8)
     }
 
     public func createDirs(at path: URL) throws {
+        if !path.isFileURL {
+            throw FileUtilsErrors.invalidFileURL(url: path)
+        }
         let fm = FileManager.default
         try fm.createDirectory(at: path, withIntermediateDirectories: true, attributes: nil)
     }
 
     public func delete(at path: URL) throws {
+        if !path.isFileURL {
+            throw FileUtilsErrors.invalidFileURL(url: path)
+        }
         let fm = FileManager.default
         try fm.removeItem(at: path)
     }
